+(*)Specify sponge damping rates in [T-1]
  Specify sponge damping rates in [T-1], instead of [s-1], for code
simplification and expanded dimensional consistency testing.  This required the
addition of a unit_scale_type argument to two initialization routines, and an
obvious bug was fixed in BFB_initialize_sponges_southonly, which is not actively
tested.  All answers are bitwise identical in the MOM6-examples test cases, but
some interfaces changed and there could be answer changes in one user-provided
example.

Conflicts:
	src/user/dumbbell_initialization.F90