Explicit array rotation index; modulo chksum turns
Array index assignment in rotation is now explicit (i.e. A(:,:)).

Modulo operators are applied to the turns in the rotated mpp checksums,
in order to prevent redundant allocations of identical arrays.  Explicit
deallocation of the rotated checksum has also been added.

An error in the comments of external forcing diagnostics was also
amended.

Thanks to Robert Hallberg for these suggestions.